+++ src/dddmr_lego_loam/lego_loam_bor/test/test_node/mapping_trt_test_node.cpp
@@ -26,6 +26,9 @@ public:
         this->create_publisher<sensor_msgs::msg::PointCloud2>("person_point_cloud", 1);  
     golden_point_cloud_pub_ = 
         this->create_publisher<sensor_msgs::msg::PointCloud2>("golden_point_cloud2", 1);  
+    sub_playing_state_ = this->create_subscription<std_msgs::msg::Bool>(
+        "is_bag_playing", 1,
+        std::bind(&MappingTestNode::playingStateCb, this, std::placeholders::_1));
     timer_ = this->create_wall_timer(
         100ms, std::bind(&MappingTestNode::testCb, this));
     RCLCPP_INFO(this->get_logger(), "MappingTestNode: %s has been started.", this->get_name());
@@ -35,7 +38,7 @@ public:
   bool is_passed_ = false;
 
 private:
-  OptimizedICPGN icp_opti_;
+  bool is_playing_ = false;
   enum class State {
     TRIGGER_MAPPING,
     WAIT_MAPPING_DONE,
@@ -49,11 +52,18 @@ private:
     switch (current_state_) {
     case State::TRIGGER_MAPPING: {
       std_msgs::msg::Bool pause_mapping;
+      pause_mapping.data = false;
       trigger_bag_mapping_pub_->publish(pause_mapping);
-      std::filesystem::remove_all("/tmp/testing_pg");
-      RCLCPP_INFO(this->get_logger(), "Changing state to WAIT_MAPPING_DONE");
-      current_state_ = State::WAIT_MAPPING_DONE;
-      break;
+      if(!is_playing_){
+        break;
+      }
+      else{
+        std::filesystem::remove_all("/tmp/testing_pg");
+        RCLCPP_INFO(this->get_logger(), "Changing state to WAIT_MAPPING_DONE");
+        current_state_ = State::WAIT_MAPPING_DONE;
+        break;
+      }
+
     }
     case State::WAIT_MAPPING_DONE:
       // Implement dir scan mechanism
@@ -165,16 +175,14 @@ private:
     }
     case State::SUCCEED: {
       RCLCPP_INFO_THROTTLE(this->get_logger(), *this->get_clock(), 1000,
-                           "Mapping test SUCCEEDED! ICP score: %.2f",
-                           icp_opti_.GetFitnessScore());
+                           "Mapping trt test SUCCEEDED!");
       is_done_ = true;
       is_passed_ = true;
       break;
     }
     case State::FAIL: {
       RCLCPP_ERROR_THROTTLE(this->get_logger(), *this->get_clock(), 1000,
-                            "Mapping test FAILED! ICP score: %.2f",
-                            icp_opti_.GetFitnessScore());
+                            "Mapping trt test FAILED!");
       is_done_ = true;
       is_passed_ = false;
       break;
@@ -185,6 +193,10 @@ private:
   rclcpp::Publisher<std_msgs::msg::Bool>::SharedPtr trigger_bag_mapping_pub_;
   rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr person_point_cloud_pub_;
   rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr golden_point_cloud_pub_;
+  rclcpp::Subscription<std_msgs::msg::Bool>::SharedPtr sub_playing_state_;
+  void playingStateCb(const std_msgs::msg::Bool::SharedPtr msg){
+    is_playing_ = msg->data;
+  }
   rclcpp::TimerBase::SharedPtr timer_;
   State current_state_ = State::TRIGGER_MAPPING;
 };
